+++ packages/agent/src/subagents/__init__.py
@@ -1,1 +1,4 @@
-# Subagent definitions will be added in Tasks 6-8
+from .copywriter import create_copywriter
+from .director import create_director
+from .researcher import create_researcher
+from .sound_engineer import create_sound_engineer